refactored PrivateRoute and debugged App
