+++ essentialism/src/components/PrivateRoute.js
@@ -6,18 +6,33 @@ import { Context } from '../Context';
 
 const PrivateRoute = ({
     component: Component,
+    token,
+    errorStatusCode,
     ...rest
 }) => {
     const { authenticated } = useContext(Context);
     return (
         <Route
             {...rest}
-            render={ props => (authenticated ?
-                <Component {...props} /> :
-                <Redirect to='/login' />)
+            render={ props =>
+                token && errorStatusCode !== 403 ? (
+                <Component {...props} /> 
+                ) : (
+                <Redirect to='/login' />
+                )
             }
         />
     );
 };
 
-export default PrivateRoute;
+const mapStateToProps = ({ token, errorStatusCode }) => ({
+    errorStatusCode,
+    token
+});
+
+export default withRouter(
+    connect(
+        mapStateToProps,
+        {}
+    )(PrivateRoute)
+);
